backend/pyproject.toml
[project]
name = "grants-management-backend"
version = "0.1.0"
description = "Backend API for Grants Management System"
requires-python = ">=3.11"
dependencies = [
    "django>=4.2,<5.0",
    "djangorestframework>=3.14.0",
    "psycopg2-binary>=2.9.0",
    "python-decouple>=3.8",
    "djangorestframework-simplejwt>=5.3.0",
    "django-cors-headers>=4.3.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "django-extensions>=3.2.0",
    "pillow>=10.0.0",
    "python-multipart>=0.0.6",
    "drf-nested-routers>=0.93.4",
    "bleach>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-django>=4.5.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "factory-boy>=3.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3